Reverted variable changes, then renamed RobotClimb to ChallengeScale
Starting over on the variable changes.

Steps to reproduce - in the Form1.cs file, replace the basic array
robotClimb with challengeScale, replace the command RobotClimbPlus with
ChallengeScalePlus, replace the control in uppercase ROBOTCLIMBPLUS with
CHALLENGESCALEPLUS.  Then go to the form designer and look at properties
of where it was displaying and change all six of the labels to the new
names (lblRobotClimb -> lblChallengeScale, lblRobotClimb2 ->
lblChallengeScale2...)  This autochanges the definitions in
Form1Designer.cs and Form1.cs.  Then go to the spot where the value is
calculated and fix the logic.