Half-fixed conversion of Pyramid to Defense1 and addition of auto-mode commands
Pyramid only existed in teleop, so I had to add it to auto.  Did all the
usual renames, plus added regular variables for the auto mode collection
to match Defense2.  Added lbl boxes on the Design.

In each of the UpdateController instances: Adjusted their teleop/auto
visibility and added converters to take the calculated value, make it a
string, and jam it into the lbl.

Next step - correct Form1 from GameCommands down